requirements.txt
pytmx
pygame-ce
pyparsing  # only for SNEK lang